+++ Inkybot/Exceptions/SinkIncorrectException.cs
@@ -2,7 +2,7 @@ namespace Inkybot.Exceptions
 {
     public class SinkIncorrectException : MagingException
     {
-        public SinkIncorrectException(float sinkChange, float sink, float attemptedSinkChange) :
+        public SinkIncorrectException(decimal sinkChange, decimal sink, decimal attemptedSinkChange) :
             base($"Calculated sink was incorrect: Current sink: {sink}, sink change: {sinkChange}, attempted sink change: {attemptedSinkChange}") {
         }
     }
