+++ Inkybot/Exceptions/SinkNegativeException.cs
@@ -2,7 +2,7 @@ namespace Inkybot.Exceptions
 {
     public class SinkNegativeException : MagingException
     {
-        public SinkNegativeException(float sink) : base($"Calculated sink is lower than 0: {sink}") {
+        public SinkNegativeException(decimal sink) : base($"Calculated sink is lower than 0: {sink}") {
         }
     }
 }
